修改文件创建权限的问题

+++ msit/components/llm/msit_llm/bc_analyze/synthezier.py
@@ -150,6 +150,10 @@ class Synthesizer(object):
         env['MSIT_TEMP_DIR_NAME'] = temp_dir_name
 
         split_command = shlex.split(command)
+        if not split_command:
+            logger.error("Command is empty,please try to run it first")
+            raise ValueError()
+
 
         known_invalid_command = {
             "rm", "mv", "mkfs", "dd",
